requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
requests>=2.31.0
tqdm>=4.66.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
PyMuPDF>=1.23.0
python-docx>=1.0.0
markdown>=3.5.0